+++ py_meta_utils.py
@@ -7,10 +7,28 @@ _missing = type('_missing', (), {'__bool__': lambda x: False})()
 ABSTRACT_ATTR = '__abstract__'
 META_OPTIONS_FACTORY_CLASS_ATTR_NAME = '_meta_options_factory_class'
 
+
 McsInitArgs = namedtuple('McsInitArgs', ('cls', 'name', 'bases', 'clsdict'))
 
 
 class McsArgs:
+    """
+    Data holder for the parameters to ``type.__new__()``::
+
+        class Metaclass(type):
+            def __new__(mcs, name, bases, clsdict):
+                mcs_args = McsArgs(mcs, name, bases, clsdict)
+                # do stuff
+                return super().__new__(*mcs_args)
+
+        # or in short-hand:
+
+        class Metaclass(type):
+            def __new__(mcs, *args):
+                mcs_args = McsArgs(mcs, *args)
+                # do stuff
+                return super().__new__(*mcs_args)
+    """
     def __init__(self,
                  mcs: type,
                  name: str,
@@ -23,16 +41,28 @@ class McsArgs:
 
     @property
     def module(self) -> Union[str, None]:
+        """
+        Returns the module of the class-under-construction, or ``None``.
+        """
         return self.clsdict.get('__module__')
 
     @property
     def qualname(self) -> str:
+        """
+        Returns the fully qualified name of the class, if possible, otherwise just the
+        class name.
+        """
         if self.module:
             return self.module + '.' + self.name
         return self.name
 
     @property
     def Meta(self) -> object:
+        """
+        Returns the class ``Meta`` from the class-under-construction.
+
+        Raises ``KeyError`` if it's not present.
+        """
         return self.clsdict['Meta']
 
     @property
@@ -52,18 +82,21 @@ class McsArgs:
 
 
 class MetaOption:
+    """
+    Base class for custom meta options.
+    """
     def __init__(self, name: str, default: Any = None, inherit: bool = False):
         self.name = name
         self.default = default
         self.inherit = inherit
 
-    def get_value(self, Meta: object, base_classes_meta, mcs_args: McsArgs):
+    def get_value(self, Meta: object, base_classes_meta, mcs_args: McsArgs) -> Any:
         """
-        :param Meta: the class Meta (if any) from the class (NOTE: this will
-                     be a plain object, NOT an instance of MetaOptionsFactory)
-        :param base_classes_meta: the MetaOptionsFactory (if any) from the
-                                  base class of the class
-        :param mcs_args: the McsArgs for the class
+        :param Meta: the class Meta (if any) from the class-under-construction (NOTE:
+                     this is a plain object, NOT an instance of MetaOptionsFactory)
+        :param base_classes_meta: the MetaOptionsFactory instance (if any) from the
+                                  base class of the class-under-construction
+        :param mcs_args: the McsArgs for the class-under-construction
         """
         value = self.default
         if self.inherit and base_classes_meta is not None:
@@ -73,9 +106,17 @@ class MetaOption:
         return value
 
     def check_value(self, value: Any, mcs_args: McsArgs):
+        """
+        Optional callback to verify the user provided a valid value.
+
+        Your implementation should assert/raise with an error message if invalid.
+        """
         pass
 
     def contribute_to_class(self, mcs_args: McsArgs, value: Any):
+        """
+        Optional callback to modify the :class:`McsArgs` of the class-under-construction.
+        """
         pass
 
     def __repr__(self):
@@ -87,6 +128,21 @@ class MetaOption:
 
 
 class AbstractMetaOption(MetaOption):
+    """
+    A meta option that allows designating a class as abstract, using either::
+
+        class SomeAbstractBase(metaclass=MetaclassWithAnOptionsFactory):
+            __abstract__ = True
+
+        # or
+
+        class SomeAbstractBase(metaclass=MetaclassWithAnOptionsFactory):
+            class Meta:
+                abstract = True
+
+    In the latter case, we make sure to set the ``__abstract__`` class attribute
+    for backwards compatibility with libraries that do not understand Meta options.
+    """
     def __init__(self):
         super().__init__(name='abstract', default=False, inherit=False)
 
@@ -117,16 +173,44 @@ class EnsureProtected(type):
 
 
 class MetaOptionsFactory(metaclass=EnsureProtected):
+    """
+    Base class for meta options factory classes. Subclasses should either set
+    :attr:`_options` to a list of :class:`MetaOption` subclasses (or instances)::
+
+        class MyMetaOptionsFactory(MetaOptionsFactory):
+            _options = [AbstractMetaOption]
+
+    Or override :meth:`_get_meta_options` to return a list of :class:`MetaOption`
+    *instances*::
+
+        class MyMetaOptionsFactory(MetaOptionsFactory):
+            def _get_meta_options(self):
+                return [AbstractMetaOption()]
+
+    IMPORTANT: If you add any attributes and/or methods to your factory subclass,
+    they *must* be protected (ie, prefixed with an `_` character).
+    """
     _options = []
 
     def __init__(self):
         self._mcs_args: McsArgs = None
 
     def _get_meta_options(self) -> List[MetaOption]:
+        """
+        Returns a list of :class:`MetaOption` instances that this factory supports.
+        """
         return [option if isinstance(option, MetaOption) else option()
                 for option in self._options]
 
     def _contribute_to_class(self, mcs_args: McsArgs):
+        """
+        Where the magic happens. Takes one parameter, the :class:`McsArgs` of the
+        class-under-construction, and processes the declared ``class Meta`` from
+        it (if any). We fill ourself with the declared meta options' name/value pairs,
+        give the declared meta options a chance to also contribute to the class-under-
+        construction, and finally replace the class-under-construction's ``class Meta``
+        with this populated factory instance (aka ``self``).
+        """
         self._mcs_args = mcs_args
 
         Meta = mcs_args.clsdict.pop('Meta', None)
@@ -143,6 +227,12 @@ class MetaOptionsFactory(metaclass=EnsureProtected):
             option.contribute_to_class(mcs_args, option_value)
 
     def _fill_from_meta(self, Meta: object, base_classes_meta, mcs_args: McsArgs):
+        """
+        Iterate over our supported meta options, and set attributes on the factory
+        instance (self) for each meta option's name/value. Raises ``TypeError`` if
+        we discover any unsupported meta options on the class-under-construction's
+        ``class Meta``.
+        """
         # Exclude private/protected fields from the Meta
         meta_attrs = {} if not Meta else {k: v for k, v in vars(Meta).items()
                                           if not k.startswith('_')}
@@ -164,7 +254,7 @@ class MetaOptionsFactory(metaclass=EnsureProtected):
                     attrs=', '.join(sorted(meta_attrs.keys()))))
 
     def __repr__(self):
-        return '<{cls} meta_options={attrs!r}>'.format(
+        return '{cls}(options={attrs!r})'.format(
             cls=self.__class__.__name__,
             attrs={option.name: getattr(self, option.name, None)
                    for option in self._get_meta_options()})
@@ -173,7 +263,61 @@ class MetaOptionsFactory(metaclass=EnsureProtected):
 def apply_factory_meta_options(
         mcs_args: McsArgs,
         default_factory_class: Type[MetaOptionsFactory] = MetaOptionsFactory,
-        factory_attr_name: Optional[str] = None) -> MetaOptionsFactory:
+        factory_attr_name: str = META_OPTIONS_FACTORY_CLASS_ATTR_NAME) \
+        -> MetaOptionsFactory:
+    """
+    Main entry point for consumer metaclasses. Usage::
+
+        from py_meta_utils import (AbstractMetaOption, McsArgs, MetaOptionsFactory,
+                                   apply_factory_meta_options)
+
+
+        class YourMetaOptionsFactory(MetaOptionsFactory):
+            _options = [AbstractMetaOption]
+
+
+        class YourMetaclass(type):
+            def __new__(mcs, name, bases, clsdict):
+                mcs_args = McsArgs(mcs, name, bases, clsdict)
+                apply_factory_meta_options(mcs_args, YourMetaOptionsFactory)
+                return super().__new__(*mcs_args)
+
+
+        class YourClass(metaclass=YourMetaclass):
+            pass
+
+    Subclasses of ``YourClass`` may set their ``_meta_options_factory_class``
+    attribute to a subclass of ``YourMetaOptionsFactory`` to customize
+    their own supported meta options::
+
+        from py_meta_utils import MetaOption
+
+
+        class FooMetaOption(MetaOption):
+            def __init__(self):
+                super().__init__(name='foo', default=None, inherit=True)
+
+
+        class FooMetaOptionsFactory(YourMetaOptionsFactory):
+            _options = YourMetaOptionsFactory._options + [
+                FooMetaOption,
+            ]
+
+
+        class FooClass(YourClass):
+            _meta_options_factory_class = FooMetaOptionsFactory
+
+            class Meta:
+                foo = 'bar'
+
+    :param mcs_args: The :class:`McsArgs` for the class-under-construction
+    :param default_factory_class: The default MetaOptionsFactory class to use, if
+                                  the ``factory_attr_name`` attribute is not set on
+                                  the class-under-construction
+    :param factory_attr_name: The attribute name to look for an overridden factory
+                              meta options class on the class-under-construction
+    :return: The populated instance of the factory class
+    """
     factory_cls = deep_getattr(
         mcs_args.clsdict, mcs_args.bases,
         factory_attr_name or META_OPTIONS_FACTORY_CLASS_ATTR_NAME,
@@ -184,6 +328,22 @@ def apply_factory_meta_options(
 
 
 class Singleton(type):
+    """
+    A metaclass that makes a consumer class a singleton::
+
+        from py_meta_utils import Singleton
+
+        class Foo(metaclass=Singleton):
+            pass
+
+        foo = Foo()
+        assert foo == Foo() == Foo()  # True
+
+    Note that if you subclass Foo, your subclass will also be a singleton, however,
+    a separate one from Foo. If instead you always want the most-derived subclass,
+    you should use :class:`SubclassableSingleton`.
+    """
+
     _instances = {}
 
     def __call__(cls, *args, **kwargs):
@@ -193,6 +353,26 @@ class Singleton(type):
 
 
 class SubclassableSingleton(Singleton):
+    """
+    A metaclass that makes a consumer class (and any subclasses of it) a singleton,
+    always returning the most-derived subclass. For example::
+
+        from py_meta_utils import SubclassableSingleton
+
+        class Foo(metaclass=SubclassableSingleton):
+            pass
+
+        class Bar(Foo):
+            pass
+
+        foo = Foo()
+        bar = Bar()
+        assert foo == bar == Foo() == Bar()  # True
+
+    Note that in practice, any subclasses must be imported *before* any calls to
+    the base class(es) are made, otherwise you will not get the correct subclass.
+    """
+
     def __call__(cls, *args, **kwargs):
         while cls.__subclasses__():
             cls = cls.__subclasses__()[0]
@@ -208,7 +388,8 @@ def deep_getattr(clsdict: Dict[str, Any],
     on the pre-class-construction class dictionary and base classes. In other words,
     first we look for the attribute in the class dictionary, and then we search all the
     base classes (in method resolution order), finally returning the default value if
-    the attribute was not found in any of the class dictionary or base classes.
+    the attribute was not found in any of the class dictionary or base classes (or it
+    raises `AttributeError` if no default was given).
     """
     value = clsdict.get(name, _missing)
     if value != _missing:
@@ -241,8 +422,8 @@ class OptionalMetaclass(type):
     def __setattr__(self, key, value):
         pass
 
-    def __call__(self, *args, **kwargs):
-        return self.__optional_class
+    def __call__(cls, *args, **kwargs):
+        return cls.__optional_class
 
     def __getitem__(self, item):
         return self.__optional_class
@@ -254,17 +435,15 @@ class OptionalMetaclass(type):
 class OptionalClass(metaclass=OptionalMetaclass):
     """
     Use this as a generic base class if you have classes that depend on an
-    optional package. For example, if you want to define a serializer but not
-    depend on flask_api_bundle, you should do something like this::
+    optional package. For example::
 
         try:
-            from flask_api_bundle import ma
+            from optional_dependency import SomeClass
         except ImportError:
-            from flask_unchained import OptionalClass as ma
+            from py_meta_utils import OptionalClass as SomeClass
 
-        class MySerializer(ma.ModelSerializer):
-            class Meta:
-                model = 'MyModel'
+        class Optional(SomeClass):
+            pass
     """
     def __init__(self, *args, **kwargs):
         pass